requirements.txt
# Oxyde ORM (installed separately via editable install from local path)
oxyde>=0.2.0

# Database drivers
asyncpg>=0.31.0
psycopg[binary]>=3.2.0
psycopg2-binary>=2.9.11
aiosqlite>=0.21.0,<0.22.0
aiomysql>=0.3.2
mysqlclient>=2.2.0

# ORMs
anyio>=4.0.0  # required by tortoise-orm 0.25.2 (not declared in their deps)
django>=5.1
sqlalchemy[asyncio]>=2.0.36
tortoise-orm>=0.25.1
piccolo>=1.28.0
peewee>=3.17.8
sqlmodel>=0.0.22
greenlet>=3.1.0

# Utils
orjson>=3.10.0
rich>=13.9.0
typer>=0.15.0
python-dateutil>=2.9.0
psutil>=6.1.0

# Testing/Dev
pytest>=8.3
pytest-asyncio>=0.24
